requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
beautifulsoup4==4.12.2
openai==1.12.0
python-dotenv==1.0.1
python-decouple==3.8
PyJWT==2.8.0
